Move miscellaneous components into Misc directory
