+++ apps/smart-forms-react-app/src/components/Questionnaires/QuestionnaireListToolbar.tsx
@@ -1,6 +1,6 @@
 import { Box, IconButton, InputAdornment, Tooltip, Typography } from '@mui/material';
 import React, { ChangeEvent, useState } from 'react';
-import Iconify from '../Iconify';
+import Iconify from '../Misc/Iconify';
 import { StyledRoot, StyledSearch } from './QuestionnaireListToolbar.styles';
 import { QuestionnaireListItem } from '../../interfaces/Interfaces';
 import { StyledAlert } from '../StyledComponents/Nav.styles';
