+++ apps/smart-forms-react-app/src/components/StyledComponents/Utils.styles.ts
